renamed jams.pyjams to jams.core
